chore: change virus severity value type

+++ root/shared/src/main/scala/it/unibo/pps/control/loader/configuration/dsl/VirusDSL.scala
@@ -21,5 +21,5 @@ object VirusDSL:
     def stdDevPositivityDays(stdDev: Double): Virus =
       virus.focus(_.stdDevPositivityDays).replace(stdDev)
 
-    def severeDeseaseProbability(probability: Int): Virus =
+    def severeDeseaseProbability(probability: Double): Virus =
       virus.focus(_.severeDeseaseProbability).replace(probability)
